refactor: merge legacy platform fetcher shim
