+++ tests/test_platform_fetcher.py
@@ -1,8 +1,7 @@
 import time
 from unittest.mock import MagicMock, patch
 
-from platform_fetcher import run_fetch_jobs
-from app.platforms.fetchers import fetch_atcoder
+from app.platforms.fetchers import fetch_atcoder, run_fetch_jobs
 
 
 def test_fetch_atcoder_returns_total_on_success():
